backend/requirements.txt
arxiv>=2.3.1
chromadb>=1.3.5
faiss-cpu>=1.13.0
fastapi>=0.121.2
groq>=0.34.1
habanero>=2.3.0
langchain>=1.0.7
langchain-chroma>=1.0.0
langchain-classic>=1.0.0
langchain-community>=0.4.1
langchain-core>=1.0.5
langchain-groq>=1.0.1
langchain-huggingface>=1.0.1
langchain-text-splitters>=1.0.0
pip>=25.3
python-dotenv
scholarly>=1.7.11
sentence-transformers>=5.1.2
tiktoken>=0.12.0
django
django-cors-headers